renaming git issues
