+++ src/pages/Dashboard/Dashboard.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { RouteComponentProps } from "react-router-dom";
-import Wrapper from "../../containers/Layout/PageWrapper";
+import Wrapper from "../../containers/Layouts/PageWrapper";
 import { Alert, PageSection, PageSectionVariants } from "@patternfly/react-core";
 
 type AllProps = RouteComponentProps;
